# docs/swagger.yaml
basePath: /
definitions:
  github_com_Its-Ameekh_school_software_backend_internal_apierrors.ErrorResponse:
    properties:
      code:
        description: machine-readable, e.g. "FORBIDDEN"
        type: string
      details:
        description: optional extra context, dev-facing
        type: string
      error:
        type: string
    type: object
...
  github_com_Its-Ameekh_school_software_backend_internal_models.LeaveRequest:
    properties:
      createdAt:
        type: string
      date:
        type: string
      id:
        type: integer
      reason:
        type: string
      requestedBy:
        description: parent user_id
        type: integer
      reviewedAt:
        type: string
      reviewedBy:
        type: integer
      status:
        description: PENDING | APPROVED | REJECTED
        type: string
      studentID:
        type: integer
    type: object
...
  github_com_Its-Ameekh_school_software_backend_internal_models.TeacherLeaveRequest:
    properties:
      createdAt:
        type: string
      fromDate:
        type: string
      id:
        type: integer
      leaveType:
        description: Casual | Sick Medical | Emergency Context | Personal Assignment
        type: string
      reason:
        type: string
      reviewedAt:
        type: string
      reviewedBy:
        type: integer
      status:
        description: PENDING | APPROVED | REJECTED
        type: string
      teacherID:
        type: integer
      toDate:
        type: string
    type: object
...
  internal_handlers.CreateStudentLeaveRequestBody:
    properties:
      date:
        description: '"YYYY-MM-DD"'
        type: string
      reason:
        type: string
    required:
    - date
    - reason
    type: object
...
  internal_handlers.CreateTeacherLeaveRequestBody:
    properties:
      from_date:
        description: '"YYYY-MM-DD"'
        type: string
      leave_type:
        description: Casual | Sick Medical | etc.
        type: string
      reason:
        type: string
      to_date:
        description: '"YYYY-MM-DD"'
        type: string
    required:
    - from_date
    - leave_type
    - reason
    - to_date
    type: object
...
  internal_handlers.UpdateLeaveStatusRequest:
    properties:
      status:
        description: APPROVED | REJECTED
        type: string
    required:
    - status
    type: object
...
info:
  contact: {}
  description: Backend API for the School Software platform (Preschool Management
    System).
  title: School Software API
  version: "1.0"
paths:
...
  /api/leaves/student/{id}/status:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Leave Request ID
        in: path
        name: id
        required: true
        type: integer
      - description: Status update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.UpdateLeaveStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Its-Ameekh_school_software_backend_internal_models.LeaveRequest'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Its-Ameekh_school_software_backend_internal_apierrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Its-Ameekh_school_software_backend_internal_apierrors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Its-Ameekh_school_software_backend_internal_apierrors.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_Its-Ameekh_school_software_backend_internal_apierrors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update student leave request status
      tags:
      - leave
  /api/leaves/teacher/{id}/status:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Leave Request ID
        in: path
        name: id
        required: true
        type: integer
      - description: Status update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.UpdateLeaveStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Its-Ameekh_school_software_backend_internal_models.TeacherLeaveRequest'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Its-Ameekh_school_software_backend_internal_apierrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Its-Ameekh_school_software_backend_internal_apierrors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Its-Ameekh_school_software_backend_internal_apierrors.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_Its-Ameekh_school_software_backend_internal_apierrors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update teacher leave request status
      tags:
      - leave
...
  /api/students/{id}/leave:
    get:
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_Its-Ameekh_school_software_backend_internal_models.LeaveRequest'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Its-Ameekh_school_software_backend_internal_apierrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Its-Ameekh_school_software_backend_internal_apierrors.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Its-Ameekh_school_software_backend_internal_apierrors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get student leave history
      tags:
      - leave
    post:
      consumes:
      - application/json
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      - description: Leave request payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.CreateStudentLeaveRequestBody'
      produces:
      - application/json
      responses:
        "21":
          description: ""
          schema:
            type: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Its-Ameekh_school_software_backend_internal_apierrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Its-Ameekh_school_software_backend_internal_apierrors.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Its-Ameekh_school_software_backend_internal_apierrors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a student leave request
      tags:
      - leave
...
  /api/teachers/leave:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_Its-Ameekh_school_software_backend_internal_models.TeacherLeaveRequest'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Its-Ameekh_school_software_backend_internal_apierrors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get logged-in teacher's leave history
      tags:
      - leave
    post:
      consumes:
      - application/json
      parameters:
      - description: Leave request payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.CreateTeacherLeaveRequestBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_Its-Ameekh_school_software_backend_internal_models.TeacherLeaveRequest'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Its-Ameekh_school_software_backend_internal_apierrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Its-Ameekh_school_software_backend_internal_apierrors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a teacher leave request
      tags:
      - leave
swagger: "2.0"